Refactor API
This update adds support for more correct GitHub sluggs, for example,
by adding support for duplicate headings, by depending on
`github-slugger` instead of other libraries.

This comes at the cost of a change in the API design.  Where it was
previously possible to use a different library instead of GitHub,
this feature is now removed.  Luckily, this does not harm npm-style
headings as npm has switched to GitHub style slugs (with the
exception of GitHub emoji shortcodes support).